+++ lib/data/model/product/product_card.dart
@@ -17,6 +17,15 @@ class ProductCard {
       required this.thumbnail_image,
       this.is_negotiable});
 
+  factory ProductCard.fromJson(Map<String, dynamic> json) {
+    return ProductCard(
+      product_id: json['productId'],
+      name: json['productName'],
+      price: json['price'],
+      thumbnail_image: json['thumbnailImage'],
+      is_negotiable: json['isNegotiable'] ?? '',
+    );
+  }
   @override
   String toString() {
     return 'ProductCard{product_id: $product_id, name: $name, price: $price, thumbnail_image: $thumbnail_image, is_negotiable: $is_negotiable}';
