+++ lib/ui/pages/chat/list/chat_list_page_view_model.dart
@@ -1,33 +1,20 @@
-import 'package:riverpod/riverpod.dart'; // Riverpod 관련 패키지
-import 'package:applus_market/data/repository/chat/chat_repository.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart';
 import '../../../../data/model/chat/chat_room_card.dart';
+import 'package:applus_market/data/repository/chat/chat_repository.dart';
 
-class ChatListPageViewModel extends Notifier<List<ChatRoomCard>> {
+class ChatListPageViewModel extends AsyncNotifier<List<ChatRoomCard>> {
   final ChatRepository chatRepository = ChatRepository();
 
-  // 채팅방 목록을 가져오는 비동기 함수
-  Future<void> selectChatRooms(int currentUserId) async {
-    // 비동기 호출 후 채팅방 목록 업데이트
-    List<ChatRoomCard> responseBody =
-        await chatRepository.selectChatRoomCards(currentUserId);
-    state = responseBody; // 상태 업데이트
-  }
-
-  // selectedIndex 관리 메서드 (선택적으로 추가)
-  void updateSelectedIndex(int index) {
-    // 필요시 selectedIndex 상태 업데이트
-  }
-
   @override
-  List<ChatRoomCard> build() {
-    selectChatRooms(1);
-    return state; // 상태 반환
+  Future<List<ChatRoomCard>> build() async {
+    return await selectChatRooms(1); // 초기 데이터 로드
+  }
+
+  Future<List<ChatRoomCard>> selectChatRooms(int currentUserId) async {
+    return await chatRepository.selectChatRoomCards(currentUserId);
   }
 }
 
-final ChatListNotiProvider =
-    NotifierProvider<ChatListPageViewModel, List<ChatRoomCard>>(
-  () {
-    return ChatListPageViewModel();
-  },
-);
+final chatListProvider =
+    AsyncNotifierProvider<ChatListPageViewModel, List<ChatRoomCard>>(
+        () => ChatListPageViewModel());
