+++ lib/ui/pages/chat/room/chat_room_page.dart
@@ -1,0 +1,11 @@
+import 'package:applus_market/ui/pages/chat/room/widget/chat_room_body.dart';
+import 'package:flutter/material.dart';
+
+class ChatRoomPage extends StatelessWidget {
+  const ChatRoomPage({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return ChatRoomBody();
+  }
+}
